iio: trx-rf: ad9088: Add mcs_track_decimation debugfs entry and widen to u32
Add debugfs read/write support for mcs_track_decimation allowing runtime
adjustment of the MCS tracking calibration TDC decimation rate. Change
mcs_track_decimation from u16 to u32 to match the API.

+++ drivers/iio/trx-rf/ad9088/ad9088_mcs.c
@@ -264,7 +264,7 @@ end:
 	return ret;
 }
 
-int ad9088_mcs_tracking_cal_setup(struct ad9088_phy *phy, u16 mcs_track_decimation,
+int ad9088_mcs_tracking_cal_setup(struct ad9088_phy *phy, u32 mcs_track_decimation,
 				  u16 initialize_track_cal)
 {
 	adi_apollo_device_t *device = &phy->ad9088;
